style: 로그인화면, 회원가입화면 css추가

+++ src/pages/posting/Posting.js
@@ -80,14 +80,16 @@ const Posting = () => {
     fileInput.current.value = "";
   };
 
-	const onRemoveQuestionHandler = () => {
-		dispatch(asyncRemoveQuestion({
-			questionId: question.id,
-			userId: userData.id
-		}));
-		alert('글이 삭제되었습니다.');
-		navigate('/', { state: { userData: userData }});
-	}
+  const onRemoveQuestionHandler = () => {
+
+    if ( window.confirm('삭제하시겠습니까?') ){
+    dispatch(asyncRemoveQuestion({
+    questionId: question.id,
+    userId: userData.id
+    }));
+    navigate('/', { state: { userData: userData }});
+    }
+    }
 
 
   // useEffect(() => {
